show-challange-completed: add message relating to current PR count, fix view of images and PR count with the info, fix the constant inline value in index.js and used exported value
